requirements.txt
pymupdf==1.25.5
langchain==0.3.25
langchain_community==0.3.24
langchain_core==0.3.59
langchain_google_genai==2.1.4
langchain_huggingface==0.2.0
pandas==2.2.3
pymilvus==2.5.8
python-dotenv==1.1.0

